chore: move main.js to root directory
